Step 2 - 수정사항 반영 0824_01
